Moved TokenSet into it's own file.
